// packages/generation-runtime/src/contracts.d.ts
export interface ArtifactRequirement {
    type: string;
    version?: string;
    optional?: boolean;
    multiple?: boolean;
}
export interface ArtifactDeclaration {
    type: string;
    version: string;
    optional?: boolean;
    core?: boolean;
}